001_requirements.txt
numpy
scipy
scikit-learn
pandas
matplotlib
seaborn
timm
open-clip-torch
pillow
tqdm
faiss-cpu
